requirements.txt
# requirements.txt
deepgram-sdk
ollama
openai
anthropic
google-generativeai
fpdf2[markdown]
python-dotenv